Utility method rearrangement
Moved utility methods into their own package with unit tests.
