--- nustar_pysolar/tests/test_sanity.py
@@ -1,3 +1,0 @@
-def test_sanity():
-    """Check the return value of func() for an example argument."""
-    assert True
